Private types

TwoDNavierStokes

GeophysicalFlows.TwoDNavierStokes.Params — Type

Params{T}(ν, nν, μ, nμ, calcF!)

A struct containing the parameters for the two-dimensional Navier-Stokes. Included are:

ν::Any

small-scale (hyper)-viscosity coefficient

nν::Int64

(hyper)-viscosity order, nν$≥ 1$

μ::Any

large-scale (hypo)-viscosity coefficient

nμ::Int64

(hypo)-viscosity order, nμ$≤ 0$

calcF!::Function

function that calculates the Fourier transform of the forcing, $F̂$

GeophysicalFlows.TwoDNavierStokes.Vars — Type

Vars{Aphys, Atrans, F, P}(ζ, u, v, ζh, uh, vh, Fh, prevsol)

The variables for two-dimensional Navier-Stokes:

ζ

relative vorticity

u

x-component of velocity

v

y-component of velocity

ζh

Fourier transform of relative vorticity

uh

Fourier transform of x-component of velocity

vh

Fourier transform of y-component of velocity

Fh

Fourier transform of forcing

prevsol

sol at previous time-step

GeophysicalFlows.TwoDNavierStokes.DecayingVars — Type

DecayingVars(dev, grid)

Return the vars for unforced two-dimensional Navier-Stokes problem on device dev and with grid.

GeophysicalFlows.TwoDNavierStokes.ForcedVars — Type

ForcedVars(dev, grid)

Return the vars for forced two-dimensional Navier-Stokes on device dev and with grid.

GeophysicalFlows.TwoDNavierStokes.StochasticForcedVars — Type

StochasticForcedVars(dev, grid)

Return the vars for stochastically forced two-dimensional Navier-Stokes on device dev and with grid.

SingleLayerQG

GeophysicalFlows.SingleLayerQG.Params — Type

Params{T, Aphys, Atrans, ℓ}(β, deformation_radius, eta, etah, μ, ν, nν, calcF!)

A struct containing the parameters for the SingleLayerQG problem. Included are:

β::Any

planetary vorticity $y$-gradient

deformation_radius::Any

Rossby radius of deformation

eta::Any

topographic potential vorticity

etah::Any

Fourier transform of topographic potential vorticity

μ::Any

linear drag coefficient

ν::Any

small-scale (hyper)-viscosity coefficient

nν::Int64

(hyper)-viscosity order, nν$≥ 1$

calcF!::Function

function that calculates the Fourier transform of the forcing, $F̂$

GeophysicalFlows.SingleLayerQG.BarotropicQGParams — Type

BarotropicQGParams(grid::TwoDGrid, β, eta, μ, ν, nν::Int, calcF

Constructor for BarotropicQGParams (infinite Rossby radius of deformation).

GeophysicalFlows.SingleLayerQG.EquivalentBarotropicQGParams — Type

EquivalentBarotropicQGParams(grid::TwoDGrid, β, deformation_radius, eta, μ, ν, nν::Int, calcF

Constructor for EquivalentBarotropicQGParams (finite Rossby radius of deformation).

GeophysicalFlows.SingleLayerQG.Vars — Type

Vars{Aphys, Atrans, F, P}(q, ψ, u, v, qh, , ψh, uh, vh, Fh, prevsol)

The variables for SingleLayer QG:

q

relative vorticity (+ vortex stretching)

ψ

streamfunction

u

x-component of velocity

v

y-component of velocity

qh

Fourier transform of relative vorticity (+ vortex stretching)

ψh

Fourier transform of streamfunction

uh

Fourier transform of x-component of velocity

vh

Fourier transform of y-component of velocity

Fh

Fourier transform of forcing

prevsol

sol at previous time-step

GeophysicalFlows.SingleLayerQG.DecayingVars — Function

DecayingVars(dev, grid)

Return the vars for unforced single-layer QG problem on device dev and with grid

GeophysicalFlows.SingleLayerQG.ForcedVars — Type

ForcedVars(dev, grid)

Return the vars for forced single-layer QG problem on device dev and with grid.

GeophysicalFlows.SingleLayerQG.StochasticForcedVars — Type

StochasticForcedVars(dev, grid)

Return the vars for stochastically forced barotropic QG problem on device dev and with grid.

MultiLayerQG

GeophysicalFlows.MultiLayerQG.Params — Type

Params{T, Aphys3D, Aphys2D, Aphys1D, Atrans4D, Trfft}(nlayers, g, f₀, β, ρ, H, U, eta, μ, ν, nν, calcFq!, g′, Qx, Qy, S, S⁻¹, rfftplan)

A struct containing the parameters for the MultiLayerQG problem. Included are:

nlayers::Int64

number of fluid layers

g::Any

gravitational constant

f₀::Any

constant planetary vorticity

β::Any

planetary vorticity $y$-gradient

ρ::Any

array with density of each fluid layer

H::Any

array with rest height of each fluid layer

U::Any

array with imposed constant zonal flow $U(y)$ in each fluid layer

eta::Any

array containing the topographic PV

μ::Any

linear bottom drag coefficient

ν::Any

small-scale (hyper)-viscosity coefficient

nν::Int64

(hyper)-viscosity order, nν$≥ 1$

calcFq!::Function

function that calculates the Fourier transform of the forcing, $F̂$

g′::Any

array with the reduced gravity constants for each fluid interface

Qx::Any

array containing $x$-gradient of PV due to eta in each fluid layer

Qy::Any

array containing $y$-gradient of PV due to $β$, $U$, and topographic PV in each fluid layer

S::Any

array containing coeffients for getting PV from streamfunction

S⁻¹::Any

array containing coeffients for inverting PV to streamfunction

rfftplan::Any

rfft plan for FFTs

GeophysicalFlows.MultiLayerQG.SingleLayerParams — Type

SingleLayerParams{T, Aphys3D, Aphys2D, Trfft}(β, U, eta, μ, ν, nν, calcFq!, Qx, Qy, rfftplan)

A struct containing the parameters for the SingleLayerQG problem. Included are:

β::Any

planetary vorticity y-gradient

U::Any

array with imposed constant zonal flow U(y)

eta::Any

array containing topographic PV

μ::Any

linear drag coefficient

ν::Any

small-scale (hyper)-viscosity coefficient

nν::Int64

(hyper)-viscosity order, nν$≥ 1$

calcFq!::Function

function that calculates the Fourier transform of the forcing, $F̂$

Qx::Any

array containing x-gradient of PV due to eta

Qy::Any

array containing y-gradient of PV due to β, U, and eta

rfftplan::Any

rfft plan for FFTs

GeophysicalFlows.MultiLayerQG.Vars — Type

Vars{Aphys, Atrans, F, P}(q, ψ, u, v, qh, , ψh, uh, vh, Fh, prevsol)

The variables for MultiLayer QG:

q

relative vorticity + vortex stretching

ψ

streamfunction

u

x-component of velocity

v

y-component of velocity

qh

Fourier transform of relative vorticity + vortex stretching

ψh

Fourier transform of streamfunction

uh

Fourier transform of x-component of velocity

vh

Fourier transform of y-component of velocity

Fqh

Fourier transform of forcing

prevsol

sol at previous time-step

GeophysicalFlows.MultiLayerQG.DecayingVars — Type

DecayingVars(dev, grid, params)

Return the vars for unforced multi-layer QG problem with grid and params.

GeophysicalFlows.MultiLayerQG.ForcedVars — Type

ForcedVars(dev, grid, params)

Return the vars for forced multi-layer QG problem with grid and params.

GeophysicalFlows.MultiLayerQG.StochasticForcedVars — Type

StochasticForcedVars(dev, rid, params)

Return the vars for forced multi-layer QG problem with grid and params.

SurfaceQG

GeophysicalFlows.SurfaceQG.Params — Type

Params{T}(ν, nν, calcF!)

A struct containing the parameters for Surface QG dynamics. Included are:

ν::Any

buoyancy (hyper)-viscosity coefficient

nν::Int64

buoyancy (hyper)-viscosity order

calcF!::Function

function that calculates the Fourier transform of the forcing, $F̂$

GeophysicalFlows.SurfaceQG.Vars — Type

Vars{Aphys, Atrans, F, P}(b, u, v, bh, uh, vh, Fh, prevsol)

The variables for surface QG problem:

b

buoyancy

u

x-component of velocity

v

y-component of velocity

bh

Fourier transform of buoyancy

uh

Fourier transform of x-component of velocity

vh

Fourier transform of y-component of velocity

Fh

Fourier transform of forcing

prevsol

sol at previous time-step

GeophysicalFlows.SurfaceQG.DecayingVars — Type

DecayingVars(dev, grid)

Return the vars for unforced surface QG dynamics on device dev and with grid.

GeophysicalFlows.SurfaceQG.ForcedVars — Type

ForcedVars(dev, grid)

Return the vars for forced surface QG dynamics on device dev and with grid.

GeophysicalFlows.SurfaceQG.StochasticForcedVars — Type

StochasticForcedVars(dev, grid)

Return the vars for stochastically forced surface QG dynamics on device dev and with grid.

BarotropicQGQL

GeophysicalFlows.BarotropicQGQL.Params — Type

Params{T, Aphys, Atrans}(β, eta, etah, μ, ν, nν, calcF!)

A struct containing the parameters for a barotropic QL QG problem. Included are:

β::Any

planetary vorticity $y$-gradient

eta::Any

topographic potential vorticity

etah::Any

Fourier transform of topographic potential vorticity

μ::Any

linear drag coefficient

ν::Any

small-scale (hyper)-viscosity coefficient

nν::Int64

(hyper)-viscosity order, nν$≥ 1$

calcF!::Function

function that calculates the Fourier transform of the forcing, $F̂$

GeophysicalFlows.BarotropicQGQL.Vars — Type

Vars{Aphys, Atrans, F, P}(u, v, U, uzeta, vzeta, zeta, Zeta, psi, Psi, N, NZ, uh, vh, Uh, zetah, Zetah, psih, Psih, Fh, prevsol)

The variables for barotropic QL QG:

u

x-component of small-scale velocity

v

y-component of small-scale velocity

U

x-component of large-scale velocity

uzeta

small-scale u′ζ′

vzeta

small-scale v′ζ′

zeta

small-scale relative vorticity

Zeta

large-scale relative vorticity

psi

small-scale relative vorticity

Psi

large-scale relative vorticity

Nz

small-scale nonlinear term

NZ

large-scale nonlinear term

uh

Fourier transform of x-component of small-scale velocity

vh

Fourier transform of y-component of small-scale velocity

Uh

Fourier transform of x-component of large-scale velocity

zetah

Fourier transform of small-scale relative vorticity

Zetah

Fourier transform of large-scale relative vorticity

psih

Fourier transform of small-scale relative vorticity

Psih

Fourier transform of large-scale relative vorticity

Fh

Fourier transform of forcing

prevsol

sol at previous time-step

GeophysicalFlows.BarotropicQGQL.DecayingVars — Type

DecayingVars(dev, grid)

Return the vars for unforced two-dimensional quasi-linear barotropic QG problem on device dev and with grid.

GeophysicalFlows.BarotropicQGQL.ForcedVars — Type

ForcedVars(dev, grid)

Return the vars for forced two-dimensional quasi-linear barotropic QG problem on device dev and with grid.

GeophysicalFlows.BarotropicQGQL.StochasticForcedVars — Type

StochasticForcedVars(dev, grid)

Return the vars for stochastically forced two-dimensional quasi-linear barotropic QG problem on device dev and with grid.