Functions

Functions exported from GeophysicalFlows:

Functions exported from TwoDNavierStokes:

GeophysicalFlows.TwoDNavierStokes.Problem — Function

Problem(; parameters...)

Construct a 2D turbulence problem.

GeophysicalFlows.TwoDNavierStokes.energy — Method

energy(prob)

Returns the domain-averaged kinetic energy, ∫ ½(u²+v²)dxdy / (Lx Ly), for the solution in sol.

GeophysicalFlows.TwoDNavierStokes.energy_dissipation — Method

energy_dissipation(prob)

Returns the domain-averaged energy dissipation rate. nν must be >= 1.

GeophysicalFlows.TwoDNavierStokes.energy_drag — Method

energy_drag(prob)

Returns the extraction of domain-averaged energy by drag/hypodrag μ.

GeophysicalFlows.TwoDNavierStokes.energy_work — Method

energy_work(prob)
energy_work(sol, vars, grid)

Returns the domain-averaged rate of work of energy by the forcing Fh.

GeophysicalFlows.TwoDNavierStokes.enstrophy — Method

enstrophy(prob)

Returns the domain-averaged enstrophy, ∫ ½ ζ² dxdy / (Lx Ly), for the solution in sol.

GeophysicalFlows.TwoDNavierStokes.enstrophy_dissipation — Method

enstrophy_dissipation(prob)

Returns the domain-averaged enstrophy dissipation rate. nν must be >= 1.

GeophysicalFlows.TwoDNavierStokes.enstrophy_drag — Method

enstrophy_drag(prob)

Returns the extraction of domain-averaged enstrophy by drag/hypodrag μ.

GeophysicalFlows.TwoDNavierStokes.enstrophy_work — Method

enstrophy_work(prob)
enstrophy_work(sol, vars, grid)

Returns the domain-averaged rate of work of enstrophy by the forcing Fh.

GeophysicalFlows.TwoDNavierStokes.set_zeta! — Method

set_zeta!(prob, zeta)

Set the solution sol as the transform of zeta and update variables.

GeophysicalFlows.TwoDNavierStokes.updatevars! — Method

updatevars!(prob)

Update variables in vars with solution in sol.

Functions exported from BarotropicQG:

GeophysicalFlows.BarotropicQG.energy — Method

energy(prob)
energy(sol, grid)

Returns the domain-averaged kinetic energy of solution sol: ∫ ½ (u²+v²) dxdy / (Lx Ly) = ∑ ½ k² |ψ̂|² / (Lx Ly).

GeophysicalFlows.BarotropicQG.energy_dissipation — Method

energy_dissipation(prob)
energy_dissipation(sol, vars, params, grid)

Returns the domain-averaged energy dissipation rate. nν must be >= 1.

GeophysicalFlows.BarotropicQG.energy_drag — Method

energy_drag(prob)

Returns the extraction of domain-averaged energy by drag μ.

GeophysicalFlows.BarotropicQG.energy_work — Method

energy_work(prob)
energy_work(sol, vars, grid)

Returns the domain-averaged rate of work of energy by the forcing Fh.

GeophysicalFlows.BarotropicQG.enstrophy — Method

enstrophy(prob)
enstrophy(sol, grid, vars)

Returns the domain-averaged enstrophy of solution sol.

GeophysicalFlows.BarotropicQG.enstrophy_dissipation — Method

enstrophy_dissipation(prob)
enstrophy_dissipation(sol, vars, params, grid)

Returns the domain-averaged enstrophy dissipation rate. nν must be >= 1.

GeophysicalFlows.BarotropicQG.enstrophy_drag — Method

enstrophy_drag(prob)

Returns the extraction of domain-averaged enstrophy by drag/hypodrag μ.

GeophysicalFlows.BarotropicQG.enstrophy_work — Method

enstrophy_work(prob)
enstrophy_work(sol, vars, grid)

Returns the domain-averaged rate of work of enstrophy by the forcing Fh.

GeophysicalFlows.BarotropicQG.set_zeta! — Method

set_zeta!(prob, zeta)
set_zeta!(sol, vars, params, grid)

Set the solution sol as the transform of zeta and update variables vars on the grid.

GeophysicalFlows.BarotropicQG.updatevars! — Method

updatevars!(sol, vars, params, grid)

Update the variables in vars with the solution in sol.

Functions exported from BarotropicQGQL:

GeophysicalFlows.BarotropicQGQL.Problem — Function

Problem(dev=CPU(); parameters...)

Construct a BarotropicQGQL turbulence problem on device dev.

GeophysicalFlows.BarotropicQGQL.dissipation — Method

dissipation(prob)
dissipation(sol, v, p, g)

Returns the domain-averaged dissipation rate. nν must be >= 1.

GeophysicalFlows.BarotropicQGQL.drag — Method

drag(prob)
drag(sol, vars, params, grid)

Returns the extraction of domain-averaged energy by drag μ.

GeophysicalFlows.BarotropicQGQL.energy — Method

energy(sol, grid)
energy(prob)

Returns the domain-averaged kinetic energy of sol.

GeophysicalFlows.BarotropicQGQL.enstrophy — Method

enstrophy(sol, grid, vars)
enstrophy(prob)

Returns the domain-averaged enstrophy of sol.

GeophysicalFlows.BarotropicQGQL.set_zeta! — Method

set_zeta!(prob, zeta)
set_zeta!(sol, vars, grid, zeta)

Set the solution sol as the transform of zeta and update variables vars on the grid.

GeophysicalFlows.BarotropicQGQL.updatevars! — Method

updatevars!(sol, vars, params, grid)
updatevars!(prob)

Update the vars of a problem prob that has grid and params with the solution in sol.

GeophysicalFlows.BarotropicQGQL.work — Method

work(prob)
work(sol, vars, params, grid)

Returns the domain-averaged rate of work of energy by the forcing, params.Fh.

Functions exported from MultilayerQG:

GeophysicalFlows.MultilayerQG.energies — Method

energies(vars, params, grid, sol)
energies(prob)

Returns the kinetic energy of each fluid layer KE$_1, ...,$ KE$_{n}$, and the potential energy of each fluid interface PE$_{3/2}, ...,$ PE$_{n-1/2}$, where $n$ is the number of layers in the fluid. (When $n=1$, only the kinetic energy is returned.)

The kinetic energy at the $j$-th fluid layer is

\[\textrm{KE}_j = \frac{H_j}{H} \int \frac1{2} |\boldsymbol{\nabla} \psi_j|^2 \frac{\mathrm{d}^2 \boldsymbol{x}}{L_x L_y} \ , \quad j = 1, \dots, n \ ,\]

while the potential energy that corresponds to the interface $j+1/2$ (i.e., interface between the $j$-th and $(j+1)$-th fluid layer) is

\[\textrm{PE}_{j+1/2} = \int \frac1{2} \frac{f_0^2}{g'_{j+1/2}} (\psi_j - \psi_{j+1})^2 \frac{\mathrm{d}^2 \boldsymbol{x}}{L_x L_y} \ , \quad j = 1, \dots, n-1 \ .\]

GeophysicalFlows.MultilayerQG.fluxes — Method

fluxes(vars, params, grid, sol)
fluxes(prob)

Returns the lateral eddy fluxes within each fluid layer, lateralfluxes$_1,...,$lateralfluxes$_n$ and also the vertical eddy fluxes at each fluid interface verticalfluxes$_{3/2}$, ...,$verticalfluxes$_{n-1/2}, where $n$ is the number of layers in the fluid. (When $n=1$, only the lateral fluxes are returned.)

The lateral eddy fluxes whithin the $j$-th fluid layer are

\[\textrm{lateralfluxes}_j = \frac{H_j}{H} \int U_j \, \upsilon_j \, \partial_y u_j \frac{\mathrm{d}^2 \boldsymbol{x}}{L_x L_y} \ , \quad j = 1, \dots, n \ ,\]

while the vertical eddy fluxes at the $j+1/2$-th fluid interface (i.e., interface between the $j$-th and $(j+1)$-th fluid layer) are

\[\textrm{verticalfluxes}_{j+1/2} = \int \frac{f_0^2}{g'_{j+1/2} H} (U_j - U_{j+1}) \, \upsilon_{j+1} \, \psi_{j} \frac{\mathrm{d}^2 \boldsymbol{x}}{L_x L_y} \ , \quad j = 1 , \dots , n-1.\]

GeophysicalFlows.MultilayerQG.set_q! — Method

set_q!(sol, params, vars, grid, q)
set_q!(prob)

Set the solution prob.sol as the transform of q and updates variables.

GeophysicalFlows.MultilayerQG.set_ψ! — Method

set_ψ!(params, vars, grid, sol, ψ)
set_ψ!(prob)

Set the solution prob.sol to correspond to the transform of streamfunction ψ and updates variables.

GeophysicalFlows.MultilayerQG.updatevars! — Method

updatevars!(vars, params, grid, sol)
updatevars!(prob)

Update all problem variables using sol.

Functions exported from SurfaceQG:

GeophysicalFlows.SurfaceQG.Problem — Function

Problem(; parameters...)

Construct a Surface QG turbulence problem.

GeophysicalFlows.SurfaceQG.buoyancy_dissipation — Method

buoyancy_dissipation(prob)

Returns the domain-averaged dissipation rate of surface buoyancy variance due to small scale diffusion/viscosity. nν must be >= 1.

In SQG, this is identical to twice the rate of kinetic energy dissipation

GeophysicalFlows.SurfaceQG.buoyancy_variance — Method

buoyancy_variance(prob)

Returns the domain-averaged buoyancy variance. In SQG flows this is identical to the domain-averaged velocity variance (twice the kinetic energy)

GeophysicalFlows.SurfaceQG.buoyancy_work — Method

buoyancy_work(prob)
buoyancy_work(sol, vars, grid)

Returns the domain-averaged rate of work of buoyancy variance by the forcing Fh.

GeophysicalFlows.SurfaceQG.kinetic_energy — Method

kinetic_energy(prob)

Returns the domain-averaged surface kinetic energy. In SQG, this is identical to half the domain-averaged surface buoyancy variance.

GeophysicalFlows.SurfaceQG.set_b! — Method

set_b!(prob, b)

Set the solution sol as the transform of b and update all variables.

GeophysicalFlows.SurfaceQG.updatevars! — Method

updatevars!(prob)

Update variables in vars with solution in sol.