Private types

Private types in module GeophysicalFlows:

Private types in module TwoDNavierStokes:

GeophysicalFlows.TwoDNavierStokes.ForcedVars — Method

ForcedVars(dev, grid)

Returns the vars for forced two-dimensional turbulence on device dev and with grid.

GeophysicalFlows.TwoDNavierStokes.Params — Type

Params(ν, nν, μ, nμ, calcF!)

Returns the params for two-dimensional turbulence.

GeophysicalFlows.TwoDNavierStokes.StochasticForcedVars — Method

StochasticForcedVars(dev, grid)

Returns the vars for stochastically forced two-dimensional turbulence on device dev and with grid grid.

GeophysicalFlows.TwoDNavierStokes.Vars — Method

Vars(dev, grid)

Returns the vars for unforced two-dimensional turbulence on device dev and with grid.

Private types in module BarotropicQG:

GeophysicalFlows.BarotropicQG.ForcedVars — Method

ForcedVars(dev, grid)

Returns the vars for forced two-dimensional barotropic QG problem on device dev and with grid.

GeophysicalFlows.BarotropicQG.Params — Type

Params(g::TwoDGrid, β, FU, eta, μ, ν, nν, calcFU, calcFq)

Returns the params for an unforced two-dimensional barotropic QG problem.

GeophysicalFlows.BarotropicQG.Params — Method

Params(g::TwoDGrid, β, eta::Function, μ, ν, nν, calcFU, calcFq)

Constructor for Params that accepts a generating function for the topographic PV.

GeophysicalFlows.BarotropicQG.StochasticForcedVars — Method

StochasticForcedVars(dev, grid)

Returns the vars for stochastically forced two-dimensional barotropic QG problem on device dev and with grid.

GeophysicalFlows.BarotropicQG.Vars — Method

Vars(dev, grid)

Returns the vars for unforced two-dimensional barotropic QG problem on device dev and with grid

Private types in module BarotropicQGQL:

GeophysicalFlows.BarotropicQGQL.ForcedVars — Method

ForcedVars(dev, grid)

Returns the vars for forced two-dimensional barotropic QG QL problem on device dev and grid.

GeophysicalFlows.BarotropicQGQL.Params — Type

Params

A struct that contains all parameter values for a two-dimensional barotropic QG QL problem.

GeophysicalFlows.BarotropicQGQL.Params — Method

Params(grid, β, eta::Function, μ, ν, nν, calcF)

Constructor for params that accepts a generating function for the topographic PV, eta.

GeophysicalFlows.BarotropicQGQL.StochasticForcedVars — Method

StochasticForcedVars(dev, grid)

Returns the vars for stochastically forced two-dimensional barotropic QG QL problem on device dev and grid.

GeophysicalFlows.BarotropicQGQL.Vars — Type

Vars

A struct that contains all variables for a two-dimensional barotropic QG QL problem.

GeophysicalFlows.BarotropicQGQL.Vars — Method

Vars(dev, grid)

Returns the vars for unforced two-dimensional barotropic QG problem with grid.

Private types in module MultilayerQG:

GeophysicalFlows.MultilayerQG.ForcedVars — Method

ForcedVars(dev, grid, params)

Returns the vars for forced multi-layer QG problem with grid and params.

GeophysicalFlows.MultilayerQG.StochasticForcedVars — Method

StochasticForcedVars(dev, rid, params)

Returns the vars for forced multi-layer QG problem with grid and params.

GeophysicalFlows.MultilayerQG.Vars — Method

Vars(dev, grid, params)

Returns the vars for unforced multi-layer QG problem with grid and params.