Private types

TwoDNavierStokes

GeophysicalFlows.TwoDNavierStokes.Params — Type

struct Params{T} <: AbstractParams

The parameters for a two-dimensional Navier-Stokes problem:

ν::Any: small-scale (hyper)-viscosity coefficient

nν::Int64: (hyper)-viscosity order, nν$≥ 1$

μ::Any: large-scale (hypo)-viscosity coefficient

nμ::Int64: (hypo)-viscosity order, nμ$≤ 0$

calcF!::Function: function that calculates the Fourier transform of the forcing, $F̂$

GeophysicalFlows.TwoDNavierStokes.Vars — Type

struct Vars{Aphys, Atrans, F, P} <: TwoDNavierStokesVars

The variables for two-dimensional Navier-Stokes problem:

ζ: relative vorticity

u: x-component of velocity

v: y-component of velocity

ζh: Fourier transform of relative vorticity

uh: Fourier transform of $x$-component of velocity

vh: Fourier transform of $y$-component of velocity

Fh: Fourier transform of forcing

prevsol: sol at previous time-step

GeophysicalFlows.TwoDNavierStokes.DecayingVars — Type

DecayingVars(dev, grid)

Return the variables for unforced two-dimensional Navier-Stokes problem on grid.

GeophysicalFlows.TwoDNavierStokes.ForcedVars — Type

ForcedVars(grid)

Return the variables for forced two-dimensional Navier-Stokes on grid.

GeophysicalFlows.TwoDNavierStokes.StochasticForcedVars — Type

StochasticForcedVars(grid)

Return the variables for stochastically forced two-dimensional Navier-Stokes on grid.

SingleLayerQG

GeophysicalFlows.SingleLayerQG.Params — Type

struct Params{T, Aphys, Atrans, Tℓ, TU} <: SingleLayerQGParams

The parameters for the SingleLayerQG problem.

β::Any: planetary vorticity $y$-gradient

deformation_radius::Any: Rossby radius of deformation

U::Any: Background flow in $x$ direction

eta::Any: topographic potential vorticity

etah::Any: Fourier transform of topographic potential vorticity

μ::Any: linear drag coefficient

ν::Any: small-scale (hyper)-viscosity coefficient

nν::Int64: (hyper)-viscosity order, nν$≥ 1$

calcF!::Function: function that calculates the Fourier transform of the forcing, $F̂$

Qx::Any: array containing $x$-gradient of PV due to eta

Qy::Any: array containing $y$-gradient of PV due to $U$ and topographic PV

GeophysicalFlows.SingleLayerQG.BarotropicQGParams — Type

BarotropicQGParams(grid, β, U, eta, μ, ν, nν, calcF)

Return the parameters for a Barotropic QG problem (i.e., with infinite Rossby radius of deformation).

GeophysicalFlows.SingleLayerQG.EquivalentBarotropicQGParams — Type

EquivalentBarotropicQGParams(grid, deformation_radius, β, U, eta, μ, ν, nν, calcF)

Return the parameters for an Equivalent Barotropic QG problem (i.e., with finite Rossby radius of deformation).

GeophysicalFlows.SingleLayerQG.Vars — Type

struct Vars{Aphys, Atrans, F, P} <: SingleLayerQGVars

The variables for SingleLayer QG:

q: relative vorticity (+ vortex stretching)

ψ: streamfunction

u: $x$-component of velocity

v: $y$-component of velocity

qh: Fourier transform of relative vorticity (+ vortex stretching)

ψh: Fourier transform of streamfunction

uh: Fourier transform of $x$-component of velocity

vh: Fourier transform of $y$-component of velocity

Fh: Fourier transform of forcing

prevsol: sol at previous time-step

GeophysicalFlows.SingleLayerQG.DecayingVars — Function

DecayingVars(grid)

Return the variables for unforced single-layer QG problem on grid.

GeophysicalFlows.SingleLayerQG.ForcedVars — Type

ForcedVars(grid)

Return the variables for forced single-layer QG problem on grid.

GeophysicalFlows.SingleLayerQG.StochasticForcedVars — Type

StochasticForcedVars(grid)

Return the variables for stochastically forced barotropic QG problem on grid.

MultiLayerQG

GeophysicalFlows.MultiLayerQG.Params — Type

struct Params{T, Aphys3D, Aphys2D, Atrans4D, Trfft} <: AbstractParams

The parameters for the MultiLayerQG problem.

nlayers::Int64: number of fluid layers

f₀::Any: constant planetary vorticity

β::Any: planetary vorticity $y$-gradient

b::Tuple: tuple with Boussinesq buoyancy of each fluid layer

H::Tuple: tuple with rest height of each fluid layer

U::Any: array with imposed constant zonal flow $U(y)$ in each fluid layer

eta::Any: array containing the topographic PV

topographic_pv_gradient::Tuple{T, T} where T: tuple containing the $(x, y)$ components of topographic PV large-scale gradient

μ::Any: linear bottom drag coefficient

ν::Any: small-scale (hyper)-viscosity coefficient

nν::Int64: (hyper)-viscosity order, nν$≥ 1$

calcFq!::Function: function that calculates the Fourier transform of the forcing, $F̂$

g′::Tuple: tuple with the reduced gravity constants for each fluid interface

Qx::Any: array containing $x$-gradient of PV due to eta in each fluid layer

Qy::Any: array containing $y$-gradient of PV due to $β$, $U$, and topographic PV in each fluid layer

S::Any: array containing coeffients for getting PV from streamfunction

S⁻¹::Any: array containing coeffients for inverting PV to streamfunction

rfftplan::Any: rfft plan for FFTs

GeophysicalFlows.MultiLayerQG.SingleLayerParams — Type

struct SingleLayerParams{T, Aphys3D, Aphys2D, Trfft} <: AbstractParams

The parameters for the a single-layer MultiLayerQG problem.

β::Any: planetary vorticity $y$-gradient

U::Any: array with imposed constant zonal flow $U(y)$

eta::Any: array containing the periodic component of the topographic PV

topographic_pv_gradient::Tuple{T, T} where T: tuple containing the $(x, y)$ components of topographic PV large-scale gradient

μ::Any: linear drag coefficient

ν::Any: small-scale (hyper)-viscosity coefficient

nν::Int64: (hyper)-viscosity order, nν$≥ 1$

calcFq!::Function: function that calculates the Fourier transform of the forcing, $F̂$

Qx::Any: array containing $x$-gradient of PV due to topographic PV

Qy::Any: array containing $y$-gradient of PV due to $β$, $U$, and topographic PV

rfftplan::Any: rfft plan for FFTs

GeophysicalFlows.MultiLayerQG.Vars — Type

struct Vars{Aphys, Atrans, F, P} <: AbstractVars

The variables for multi-layer QG problem.

q: relative vorticity + vortex stretching

ψ: streamfunction

u: $x$-component of velocity

v: $y$-component of velocity

qh: Fourier transform of relative vorticity + vortex stretching

ψh: Fourier transform of streamfunction

uh: Fourier transform of $x$-component of velocity

vh: Fourier transform of $y$-component of velocity

Fqh: Fourier transform of forcing

prevsol: sol at previous time-step

GeophysicalFlows.MultiLayerQG.DecayingVars — Type

DecayingVars(grid, params)

Return the variables for an unforced multi-layer QG problem with grid and params.

GeophysicalFlows.MultiLayerQG.ForcedVars — Type

ForcedVars(grid, params)

Return the variables for a forced multi-layer QG problem with grid and params.

GeophysicalFlows.MultiLayerQG.StochasticForcedVars — Type

StochasticForcedVars(grid, params)

Return the variables for a forced multi-layer QG problem with grid and params.

SurfaceQG

GeophysicalFlows.SurfaceQG.Params — Type

Params{T}(ν, nν, calcF!)

A struct containing the parameters for Surface QG dynamics. Included are:

ν::Any: buoyancy (hyper)-viscosity coefficient

nν::Int64: buoyancy (hyper)-viscosity order

calcF!::Function: function that calculates the Fourier transform of the forcing, $F̂$

GeophysicalFlows.SurfaceQG.Vars — Type

Vars{Aphys, Atrans, F, P}(b, u, v, bh, uh, vh, Fh, prevsol)

The variables for surface QG problem:

b: buoyancy

u: $x$-component of velocity

v: $y$-component of velocity

bh: Fourier transform of buoyancy

uh: Fourier transform of $x$-component of velocity

vh: Fourier transform of $y$-component of velocity

Fh: Fourier transform of forcing

prevsol: sol at previous time-step

GeophysicalFlows.SurfaceQG.DecayingVars — Type

DecayingVars(grid)

Return the variables for unforced surface QG dynamics on grid.

GeophysicalFlows.SurfaceQG.ForcedVars — Type

ForcedVars(grid)

Return the variables for forced surface QG dynamics on grid.

GeophysicalFlows.SurfaceQG.StochasticForcedVars — Type

StochasticForcedVars(grid)

Return the variables for stochastically forced surface QG dynamics on grid.

BarotropicQGQL

GeophysicalFlows.BarotropicQGQL.Params — Type

Params{T, Aphys, Atrans}(β, eta, etah, μ, ν, nν, calcF!)

A struct containing the parameters for a barotropic QL QG problem. Included are:

β::Any: planetary vorticity $y$-gradient

eta::Any: topographic potential vorticity

etah::Any: Fourier transform of topographic potential vorticity

μ::Any: linear drag coefficient

ν::Any: small-scale (hyper)-viscosity coefficient

nν::Int64: (hyper)-viscosity order, nν$≥ 1$

calcF!::Function: function that calculates the Fourier transform of the forcing, $F̂$

GeophysicalFlows.BarotropicQGQL.Vars — Type

Vars{Aphys, Atrans, F, P}(u, v, U, uzeta, vzeta, zeta, Zeta, psi, Psi, N, NZ, uh, vh, Uh, zetah, Zetah, psih, Psih, Fh, prevsol)

The variables for barotropic QL QG:

u: $x$-component of small-scale velocity

v: $y$-component of small-scale velocity

U: $x$-component of large-scale velocity

uzeta: small-scale $u′ζ′$

vzeta: small-scale $v′ζ′$

zeta: small-scale relative vorticity

Zeta: large-scale relative vorticity

psi: small-scale streamfunction

Psi: large-scale streamfunction

Nz: small-scale nonlinear term

NZ: large-scale nonlinear term

uh: Fourier transform of $x$-component of small-scale velocity

vh: Fourier transform of $y$-component of small-scale velocity

Uh: Fourier transform of $x$-component of large-scale velocity

zetah: Fourier transform of small-scale relative vorticity

Zetah: Fourier transform of large-scale relative vorticity

psih: Fourier transform of small-scale relative vorticity

Psih: Fourier transform of large-scale relative vorticity

Fh: Fourier transform of forcing

prevsol: sol at previous time-step

GeophysicalFlows.BarotropicQGQL.DecayingVars — Type

DecayingVars(grid)

Return the variables for unforced two-dimensional quasi-linear barotropic QG problem on grid.

GeophysicalFlows.BarotropicQGQL.ForcedVars — Type

ForcedVars(grid)

Return the variables for forced two-dimensional quasi-linear barotropic QG problem on grid.

GeophysicalFlows.BarotropicQGQL.StochasticForcedVars — Type

StochasticForcedVars(grid)

Return the variables for stochastically forced two-dimensional quasi-linear barotropic QG problem on grid.