Functions

Functions exported from GeophysicalFlows:

Functions exported from TwoDNavierStokes:

GeophysicalFlows.TwoDNavierStokes.Problem — Method

Problem(; parameters...)

Construct a 2D turbulence problem.

GeophysicalFlows.TwoDNavierStokes.dissipation — Method

dissipation(prob)

Returns the domain-averaged dissipation rate. nν must be >= 1.

GeophysicalFlows.TwoDNavierStokes.drag — Method

drag(prob)

Returns the extraction of domain-averaged energy by drag/hypodrag μ.

GeophysicalFlows.TwoDNavierStokes.energy — Method

energy(prob)

Returns the domain-averaged kinetic energy in the Fourier-transformed vorticity solution sol.

GeophysicalFlows.TwoDNavierStokes.enstrophy — Method

enstrophy(prob)

Returns the domain-averaged enstrophy in the Fourier-transformed vorticity solution sol.

GeophysicalFlows.TwoDNavierStokes.set_zeta! — Method

set_zeta!(prob, zeta)

Set the solution sol as the transform of zeta and update variables v on the grid g.

GeophysicalFlows.TwoDNavierStokes.updatevars! — Method

updatevars!(prob)

Update the vars in v on the grid g with the solution in sol.

GeophysicalFlows.TwoDNavierStokes.work — Method

work(prob)
work(sol, v, g)

Returns the domain-averaged rate of work of energy by the forcing Fh.

Functions exported from BarotropicQG:

GeophysicalFlows.BarotropicQG.dissipation — Method

dissipation(prob)
dissipation(s, v, p, g)

Returns the domain-averaged dissipation rate. nν must be >= 1.

GeophysicalFlows.BarotropicQG.drag — Method

drag(prob)
drag(s, v, p, g)

Returns the extraction of domain-averaged energy by drag μ.

GeophysicalFlows.BarotropicQG.energy — Method

energy(sol, g)
energy(prob)

Returns the domain-averaged kinetic energy of sol.

GeophysicalFlows.BarotropicQG.enstrophy — Method

enstrophy(sol, g, v)
enstrophy(prob)

Returns the domain-averaged enstrophy of sol.

GeophysicalFlows.BarotropicQG.meanenergy — Method

meanenergy(prob)

Returns the energy of the domain-averaged U.

GeophysicalFlows.BarotropicQG.meanenstrophy — Method

meanenstrophy(prob)

Returns the enstrophy of the domain-averaged U.

GeophysicalFlows.BarotropicQG.set_zeta! — Method

set_zeta!(prob, zeta)
set_zeta!(s, v, g, zeta)

Set the solution sol as the transform of zeta and update variables v on the grid g.

GeophysicalFlows.BarotropicQG.updatevars! — Method

updatevars!(v, s, g)

Update the vars in v on the grid g with the solution in sol.

GeophysicalFlows.BarotropicQG.work — Method

work(prob)
work(s, v, p, g)

Returns the domain-averaged rate of work of energy by the forcing Fqh.

Functions exported from BarotropicQGQL:

GeophysicalFlows.BarotropicQGQL.Problem — Method

Problem(; parameters...)

Construct a BarotropicQGQL turbulence problem.

GeophysicalFlows.BarotropicQGQL.dissipation — Method

dissipation(prob)
dissipation(sol, v, p, g)

Returns the domain-averaged dissipation rate. nnu must be >= 1.

GeophysicalFlows.BarotropicQGQL.drag — Method

drag(prob)
drag(sol, v, p, g)

Returns the extraction of domain-averaged energy by drag mu.

GeophysicalFlows.BarotropicQGQL.energy — Method

energy(sol, g)
energy(prob)

Returns the domain-averaged kinetic energy of sol.

GeophysicalFlows.BarotropicQGQL.enstrophy — Method

enstrophy(sol, g, v)
enstrophy(prob)

Returns the domain-averaged enstrophy of sol.

GeophysicalFlows.BarotropicQGQL.set_zeta! — Method

set_zeta!(prob, zeta)
set_zeta!(sol, v, g, zeta)

Set the solution sol as the transform of zeta and update variables v on the grid g.

GeophysicalFlows.BarotropicQGQL.updatevars! — Method

updatevars!(v, s, g)

Update the vars in v on the grid g with the solution in sol.

GeophysicalFlows.BarotropicQGQL.work — Method

work(prob)
work(sol, v, p, g)

Returns the domain-averaged rate of work of energy by the forcing Fh.

Functions exported from MultilayerQG:

GeophysicalFlows.MultilayerQG.energies — Method

energies(prob)

Returns the kinetic energy of each fluid layer KE1,...,KEnlayers, and the potential energy of each fluid interface PE{3/2},...,PE{nlayers-1/2}.

GeophysicalFlows.MultilayerQG.fluxes — Method

fluxes(prob)

Returns the lateral eddy fluxes within each fluid layer lateralfluxes1,...,lateralfluxesnlayers and also the vertical eddy fluxes for each fluid interface verticalfluxes{3/2},...,verticalfluxes{nlayers-1/2}

GeophysicalFlows.MultilayerQG.set_psi! — Method

set_psi!(prob)

Set the solution prob.sol to correspond to a streamfunction psi and updates variables.

GeophysicalFlows.MultilayerQG.set_q! — Method

set_q!(prob)

Set the solution prob.sol as the transform of q and updates variables.

GeophysicalFlows.MultilayerQG.updatevars! — Method

updatevars!(prob)

Update prob.vars using prob.sol.